requirements.txt
discord.py
nekos.py
jishaku
NekoMimi
